test weitergefahren

+++ src/ch/bfh/advancedweb/peer2peer/test/UserCRUDTest.java
@@ -15,6 +15,7 @@ import org.junit.Test;
 import ch.bfh.advancedweb.peer2peer.model.User;
 
 public class UserCRUDTest {
+	
 
 	@Test
 	public void createUser() {
@@ -24,7 +25,7 @@ public class UserCRUDTest {
 		user.setLastname("Moser");
 		user.setBirthdate(new Date(1988,03,22));
 		user.setCity("Fraubrunnen");
-		user.setCountry("Switzerland");
+		user.setCountry("Bangladesch");
 		user.setEmail("mail.martinmoser@gmail.com");
 		user.setExpenses(1000.0);
 		user.setExsisting_credits(0);
@@ -45,10 +46,7 @@ public class UserCRUDTest {
 		em.getTransaction().commit();
 		
 	}
-	
-	
-	
-	/*
+
 	@Test
 	public void readUser(){
 		EntityManager em = Persistence.createEntityManagerFactory(
@@ -61,7 +59,7 @@ public class UserCRUDTest {
 		List<User> foundUsers = q.getResultList();
 		if(foundUsers != null){
 			User firstUser = (User)foundUsers.get(0);
-			Assert.assertTrue(firstUser.getFirstname().equals("Martin") && firstUser.getLastname().equals("Moser") && firstUser.getEmail().equals("mail.martinmoser@gmail.com"));
+			Assert.assertTrue(firstUser.getEmail().equals("mail.martinmoser@gmail.com"));
 		}
 	}
 	
@@ -77,12 +75,9 @@ public class UserCRUDTest {
 		List<User> foundUsers = q.getResultList();
 		User firstUser = foundUsers.get(0);
 
-		// Write access needs a transaction
 		em.getTransaction().begin();
 		firstUser.setFirstname("Christian");
 		em.getTransaction().commit();
-		// Entity is persisted automatically after commit because it is managed
-		// by jpa.
 
 		Assert.assertTrue(firstUser.getFirstname().equals("Christian"));
 		
@@ -91,22 +86,23 @@ public class UserCRUDTest {
 	@Test
 	public void deleteTest(){
 		
-		EntityManager em = Persistence.createEntityManagerFactory(
-				"ch.bfh.advancedweb.peer2peer.model").createEntityManager();
+		EntityManager em = Persistence.createEntityManagerFactory("ch.bfh.advancedweb.peer2peer.model").createEntityManager();
 
 		Query q = em.createQuery("select a from User a where a.email=:email");
 		q.setParameter("email", "mail.martinmoser@gmail.com");
 		@SuppressWarnings("unchecked")
 		List<User> foundUsers = q.getResultList();
 		User firstUser = foundUsers.get(0);
-
-		// Write access needs a transaction
-		em.getTransaction().begin();
-		em.remove(firstUser);
-		em.getTransaction().commit();
+		
+		if(firstUser != null){
+			
+			// Write access needs a transaction
+			em.getTransaction().begin();
+			em.remove(firstUser);
+			em.getTransaction().commit();
+			
+		}
 		
 	}
-	
-	*/
 
 }
